pyproject.toml
[project]
name = "py-noodle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "c-two>=0.2.4",
    "fastapi>=0.116.1",
    "pyyaml>=6.0.2",
    "pyzmq>=27.0.0",
]

[dependency-groups]
dev = [
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
]
